+++ packages/expo-modules-core/android/src/androidTest/java/expo/modules/kotlin/jni/types/ArrayBufferConversionTest.kt
@@ -25,6 +25,18 @@ class ArrayBufferConversionTest {
   )
 
   @Test
+  fun array_buffer_should_be_convertible_from_u8array() = conversionTest<JavaScriptArrayBuffer, _>(
+    jsValue = "new Uint8Array([0x00, 0xff])",
+    nativeAssertion = { arrayBuffer ->
+      Truth.assertThat(arrayBuffer.size()).isEqualTo(2)
+      Truth.assertThat(arrayBuffer.readByte(0)).isEqualTo(0x00.toByte())
+      Truth.assertThat(arrayBuffer.readByte(1)).isEqualTo(0xff.toByte())
+    },
+    map = {},
+    jsAssertion = {}
+  )
+
+  @Test
   fun js_array_buffer_should_be_returned() = conversionTest<JavaScriptArrayBuffer, _>(
     jsValue = "new Uint8Array([0x00, 0xff]).buffer",
     nativeAssertion = { arrayBuffer ->